Move the public key tags api test to it's own module
